+++ Modular Gameplay Overhaul/Modules/Arsenal/Patchers/Weapons/MeleeWeaponAnimateSpecialMovePatcher.cs
@@ -24,7 +24,7 @@ internal sealed class MeleeWeaponAnimateSpecialMovePatcher : HarmonyPatcher
     [HarmonyPrefix]
     private static bool MeleeWeaponAnimateSpecalMovePrefix(MeleeWeapon __instance, ref Farmer ___lastUser, Farmer who)
     {
-        if (__instance.type.Value != MeleeWeapon.stabbingSword || MeleeWeapon.attackSwordCooldown > 0)
+        if (__instance.isScythe() || __instance.type.Value != MeleeWeapon.stabbingSword || MeleeWeapon.attackSwordCooldown > 0)
         {
             return true; // run original logic
         }